Separate enum into individual classes
